migrate SimplifyComparisonArithmetics rule
